uvstream: fix alignment for leader, trailer and payload transfer sized
Fixes the alignment for leader, trailer and payload transfer.
The alignment is defined by the device and can be determined using
the SI Info field. The leader and trailer size is aligned according
to the provided value.

To align the payload size the maximum payload size defined by
the host is first aligned by 'rounding' downwards and than the
'SI Payload Transfer Size' is determined. The last step is to
also align the 'SI Final Transfer1 Size'.