introspection: workaround binding issue of boolean returning functions (#316)
The function gc_boolean_get_value take a GError ** and it returns a
boolean, but the return value doesn't indicate if the function call
failed. The result is that the automatically generated bindings assume
there is no data in the return value. In Rust, that translates to a
Result with no value.

Fix the binding by implementing functions taking an ouput parameter, and
using rename-to annotation to replace the functions designed for C use.

Fixes #315