gvdevice: improve auto packet size
The computation of the expected test packet was wrong on some platform, so don't
use sizeof.

Also, GevSCPSPacketSize increment was not correctly take into account.

And finally, don't limit the precision to 16 but to increment instead. I fear
forcing the precision may conflict with cameras with increment not a divisor of
16 (which would be surprising, but who knows (yes, the GigEVision documentation,
but we can not use it :( )).

Fix #718